changing to proper types

+++ voice/response_voice.go
@@ -18,7 +18,7 @@ type Response struct {
 	PhoneNumberSid string
 	StartTime      common.JsonTime `json:"start_time"`
 	EndTime        common.JsonTime `json:"end_time"`
-	Duration       float64         `json:"duration"`
+	Duration       int             `json:"duration"`
 	AnsweredBy     string          `json:"answered_by"`
 	ForwardedFrom  string          `json:"fowarded_from"`
 	CallerName     string          `json:"caller_name"`
